+++ src/llm/io_processing/gemma4/gemma4_tool_parser.hpp
