+++ src/llm/io_processing/output_parser.cpp
@@ -31,7 +31,7 @@
 #include "gemma4/gemma4_reasoning_parser.hpp"
 #include "gptoss/reasoning_parser.hpp"
 #include "lfm2/lfm2_tool_parser.hpp"
-#include "gemma4/tool_parser.hpp"
+#include "gemma4/gemma4_tool_parser.hpp"
 
 namespace ovms {
 OutputParser::TagLookupStatus OutputParser::StreamOutputCache::lookupTag(const std::string& tag) const {
